Move requests to linkedin api client.
